# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  villages: number;
  farms: number;
  cows: number;
  activeHealthAlerts: number;
  milkCollectedToday: number;
  pendingOrders: number;
}

export interface MilkVolumePoint {
  date: string;
  liters: number;
}

export interface Village {
  id: number;
  name: string;
  district: string;
  state: string;
}

export interface Farm {
  id: number;
  villageId: number;
  villageName: string;
  farmerName: string;
  phone: string;
  address: string;
  cowCount: number;
}

export interface Cow {
  id: number;
  farmId: number;
  farmName: string;
  tagId: string;
  breed: string;
  dateOfBirth: string;
  lactationStatus: string;
}

export interface CollectionCenter {
  id: number;
  villageId: number;
  villageName: string;
  name: string;
  location: string;
  farmCount: number;
}

export interface ProcessingPlant {
  id: number;
  name: string;
  location: string;
  batchCount: number;
}

export interface Supermarket {
  id: number;
  name: string;
  location: string;
  orderCount: number;
}

export interface FieldLog {
  id: number;
  farmId: number;
  farmName: string;
  fieldWorkerName: string;
  logDate: string;
  checkIn: string;
  /** @nullable */
  checkOut?: string | null;
  milkLiters: number;
  gheeKg: number;
  dahiKg: number;
  notes?: string;
  status: string;
}

export interface FieldLogInput {
  farmId: number;
  /** @minLength 1 */
  fieldWorkerName: string;
  logDate: string;
  /** @minimum 0 */
  milkLiters: number;
  /** @minimum 0 */
  gheeKg: number;
  /** @minimum 0 */
  dahiKg: number;
  notes?: string;
}

export interface FieldLogSummary {
  milkLiters: number;
  gheeKg: number;
  dahiKg: number;
  activeSessions: number;
  completedLogs: number;
}

export interface CctvCamera {
  id: number;
  farmId: number;
  farmName: string;
  name: string;
  location: string;
  /** @nullable */
  streamUrl?: string | null;
  status: string;
  /** @nullable */
  lastSeen?: string | null;
}

export interface CctvCameraInput {
  farmId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  location: string;
  /** @nullable */
  streamUrl?: string | null;
}