# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /field-logs:
    get:
      operationId: listFieldLogs
      tags: [field-operations]
      summary: List daily field worker production logs
      responses:
        "200":
          description: Field logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FieldLog"
    post:
      operationId: createFieldLog
      tags: [field-operations]
      summary: Start a field worker session and record production
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldLogInput"
      responses:
        "201":
          description: Created field log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldLog"
  /field-logs/{id}/checkout:
    post:
      operationId: checkoutFieldLog
      tags: [field-operations]
      summary: Close a field worker session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Closed field log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldLog"
        "404":
          description: Field log not found
...
  /cctv/cameras:
    get:
      operationId: listCctvCameras
      tags: [cctv]
      summary: List farm CCTV cameras
      responses:
        "200":
          description: CCTV cameras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CctvCamera"
    post:
      operationId: createCctvCamera
      tags: [cctv]
      summary: Add a farm CCTV camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CctvCameraInput"
      responses:
        "201":
          description: Created CCTV camera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvCamera"
components:
  schemas:
...
    FieldLog:
      type: object
      required:
        - id
        - farmId
        - farmName
        - fieldWorkerName
        - logDate
        - checkIn
        - milkLiters
        - gheeKg
        - dahiKg
        - status
      properties:
        id:
          type: number
        farmId:
          type: number
        farmName:
          type: string
        fieldWorkerName:
          type: string
        logDate:
          type: string
          format: date
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: ["string", "null"]
          format: date-time
        milkLiters:
          type: number
        gheeKg:
          type: number
        dahiKg:
          type: number
        notes:
          type: string
        status:
          type: string
    FieldLogInput:
      type: object
      required:
        - farmId
        - fieldWorkerName
        - logDate
        - milkLiters
        - gheeKg
        - dahiKg
      properties:
        farmId:
          type: number
        fieldWorkerName:
          type: string
          minLength: 1
        logDate:
          type: string
          format: date
        milkLiters:
          type: number
          minimum: 0
        gheeKg:
          type: number
          minimum: 0
        dahiKg:
          type: number
          minimum: 0
        notes:
          type: string
...
    CctvCamera:
      type: object
      required:
        - id
        - farmId
        - farmName
        - name
        - location
        - status
      properties:
        id:
          type: number
        farmId:
          type: number
        farmName:
          type: string
        name:
          type: string
        location:
          type: string
        streamUrl:
          type: ["string", "null"]
        status:
          type: string
        lastSeen:
          type: ["string", "null"]
          format: date-time
    CctvCameraInput:
      type: object
      required:
        - farmId
        - name
        - location
      properties:
        farmId:
          type: number
        name:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        streamUrl:
          type: ["string", "null"]