# lib/api-zod/src/generated/types/cctvCamera.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CctvCamera {
  id: number;
  farmId: number;
  farmName: string;
  name: string;
  location: string;
  /** @nullable */
  streamUrl?: string | null;
  status: string;
  /** @nullable */
  lastSeen?: Date | null;
}