# lib/api-zod/src/generated/types/cctvCameraInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CctvCameraInput {
  farmId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  location: string;
  /** @nullable */
  streamUrl?: string | null;
}