# lib/api-zod/src/generated/types/fieldLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FieldLog {
  id: number;
  farmId: number;
  farmName: string;
  fieldWorkerName: string;
  logDate: Date;
  checkIn: Date;
  /** @nullable */
  checkOut?: Date | null;
  milkLiters: number;
  gheeKg: number;
  dahiKg: number;
  notes?: string;
  status: string;
}