# lib/api-zod/src/generated/types/fieldLogInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FieldLogInput {
  farmId: number;
  /** @minLength 1 */
  fieldWorkerName: string;
  logDate: Date;
  /** @minimum 0 */
  milkLiters: number;
  /** @minimum 0 */
  gheeKg: number;
  /** @minimum 0 */
  dahiKg: number;
  notes?: string;
}